+++ fort-to-nite/Models/Weapons.swift
@@ -34,5 +34,4 @@ class Weapons: Decodable {
         self.group = group
         self.is_removed = is_removed
     }
-    
 }